cl/385848717: ghost: use ELIDE_PREEMPT for sync group transactions
The new kernel policy of sending TASK_PREEMPT messages on ghost-to-ghost
scheduling edges hurts the performance of any agent that schedules both its cpu
and another cpu.  The preempt from the other cpu's ghost-to-ghost edge rewakes
the agent immediately after it tried to give its core to a ghost task.

To be submitted in conjunction with kcl/395979.

Tested: api_test, edf_test