cl/385848911: ghost: add a scheduler-specific BPF program to EDF
This is just an example of how to add scheduler-specific BPF programs.  In this
case, I'm overriding the agent-independent tick program with edf-specific
programs.  We can even choose which BPF program to load at runtime, controlled
by edf_ticks_.

The example programs are just "send a tick" or "skip sending the tick".

One ugly bit is that we can't use AgentConfig.cpu_ticks_.  That controls
whether or not the scheduler-independent agent/enclave code loads the 'default'
BPF tick program from ghost.bpf.c.  Since EDF is providing its own program, it
needs to tell the libghost code to not load a program, which is done by setting
cpu_ticks_ = kAllTicks, since the kernel's default is to always send tick
messages.  That may change in the future - I'm inclined to reverse the kernel's
default.

Anyway, to get around that, we set AgentConfig.cpu_ticks_ = kAllTicks, and the
EDF-specific GlobalConfig.edf_ticks to whatever the user wants (default
kNoTicks).

One incidental change in this CL is that the EdfScheduler ctor now takes the
GlobalConfig& instead of the s32 global_cpu_id.  This way all
scheduler-specific config options get passed all the way to the scheduler's
ctor.

Tested: edf_test, agent_exp with and without --ticks