cl/385914334: Test case to trigger race between TASK_NEW and TASK_DEPARTED/TASK_AFFINITY_CHANGED.
The race is described in detail in comments above DepartedRaceAgent and
results in the agent getting a MSG_TASK_DEPARTED for a task that it has
never heard from (i.e. there is no preceding MSG_TASK_NEW for the task).

Prior to the kernel fix the test reliably fails with this signature:

    M: MSG_TASK_DEPARTED seq=1 B0/2035 on cpu 8
    third_party/ghost/tests/api_test.cc:1571: Failure
    Value of: task
    Expected: isn't equal to (nullptr)
      Actual: NULL (of type ghost::FifoTask*)

    OR

    M: MSG_TASK_AFFINITY_CHANGED seq=1 B0/2034
    third_party/ghost/tests/api_test.cc:1614: Failure
    Value of: task
    Expected: isn't equal to (nullptr)
      Actual: NULL (of type ghost::FifoTask*)

Note the 'seq=1' on MSG_TASK_DEPARTED indicating that this is the very
first message produced for the task (typically we expect seq=1 to be
associated with MSG_TASK_NEW).

TESTED=//third_party/ghost/api_test in virtme