cl/388797042: ghOSt: Get rid of open source bazel warning
For some reason, the compiler isn't smart enough to figure out that `tick_on_request` will be initialized when it is passed to `agent_bpf_init()`, so the compiler emits a warning that `tick_on_request` may be uninitialized in that call. I verified that the build fails when the switch statement does not contain a case statement for each enum class member, so it is strange that the compiler emits this warning.