cl/390199979: ghost: fix the split BPF headers problem
Consolidate the multiple copies of headers used by BPF and userspace.  No
change in functionality.

For a while, we had two copies of header files shared between BPF and
userspace.  We'd have to manually keep them in sync, which was a huge pain and
liable to cause trouble.

I forget all of the reasons why we had to have that.  One of them was that you
couldn't include stdint with BPF.  There might have been more that are no
longer a problem.

Jack suggested adding a compiler flag, but we can't easily do that to the blaze
rules.  It turns out that clang defines __BPF__ (and __bpf__) for us.  I
thought I tried those a while back, but maybe not.  Either way, it works now.

For future reference, you can tell which #defines clang sets like so:

	clang --target=bpfel -E - -dM </dev/null | grep -i bpf

Tested: edf_test and the various tools
Fixes: b/194074152