cl/390209831: ghOSt: Copybara cleanup
---

cl/390373098: ghOSt: Update WORKSPACE file to use open source ghOSt kernel repository

We have the Linux repository as a dependency so that we can use libbpf and bpftool from it to compile eBPF programs, eBPF skeletons, and eBPF userspace programs. Previously, we had the official Linux repository as a dependency since the ghOSt kernel repository was not public on GitHub, which would prevent Bazel from cloning the repository. This CL updates the dependency to use the ghOSt kernel repository instead as the ghOSt kernel repository is now public.

It is best to have the ghOSt kernel repository as a dependency as we can now potentially pull things from it into the userspace build process as needed in the future.

---

cl/390408279: ghOSt: Copybara fixes