cl/390708651: ghOSt: Fix bugs where a CpuList was moved and then the same CpuList was later accessed after the move
As noted by neelnatu@ in cl/389712071, we do not implement a move constructor for CpuList. Thus, the compiler generates a move constructor for CpuList automatically. Since the bitmap backing store for CpuList is an array that lives within the class instance, it is likely that the automatically generated move constructor copies the entire bitmap into the new CpuList. This is why we were getting away with having these bugs present in the code.

Tested: `./topology_test` and `./centralized_queuing.par ghost` on ikbe14