cl/393141460: ghost: add a runnable_timeout helper and use it in SGS
The runnable_timeout protects our app from situations where the enclave fails
to schedule a runnable task for X msec.

I'm open to having a different default value.  I went with 20s, since that
matches what we can handle in SLL.  Prior to this CL, the default was 0: the
enclave's default.

Tested: virtme and simple_exp, checked to see it worked