cl/405490273: Fix flaky ApiTest.KernelTimes test.
On debug kernels the test sporadically trips both the 100 usec 'commit_delay'
and 'switch_delay' checks. The new debug prints added to the destructor show
that the values of 'commit_delay' and 'switch_delay' are an order of magnitude
higher for the very first commit. For e.g.

switch_delays:
  28424		<< 28.4 usec
  1399		<< everything following is less than 1.5 usec.
  930
  899

commit_delays:
  185376	<< would have tripped the (commit_delay < 100 usec) check.
  3480		<< everything following is less than 3.5 usec.
  2300
  2060

I haven't been able to pinpoint exactly what config option in the DBG kernel
is responsible for the first commit discrepancy (maybe lockdep when it wants
to track a new lock order dependency?). Rather than skipping the test on dbg
kernels (prodkernel specific) or bumping up the delay bound to a ridiculous
number (e.g. 500 usec) we skip the bounds check on the various delays for the
very first commit.