cl/406186753: ghost: allow agents to change their AgentProcess's stderr FD
When working from the shell, printing to stderr from the AP is fine.  But when
the agent is launched from a daemon, we do not have visibility into anything
the AP prints.

Now, agents can set AgentConfig.stderr_fd to provide their own FD for the AP to
use for stderr.

Any ForkedProcess can use the stderr_Fd too, in case we want to catch the
stderr from a test.

Also: the agent RPC handler thread wasn't detached.  We never noticed it, but
failing to detach a thread will throw an exception when the main thread exits.
That was fun.