cl/410697037: ghOSt: Add centralized FIFO scheduler
This is a centralized FIFO scheduler that does synchronous commits. I used the SOL scheduler as a starting point.

Notes: This scheduler does not use a PrioTable which makes it easier to modify. This scheduler does not yet support switchto.

Tested: for i in {1..1000}; do ./simple_exp; done (and verified visually with SchedViz)