cl/414696186: ghost: add schedghostidle as an optional agent tracer
Additionally, we can add other BPF tools to any agent, with the family of
agent_bpf_trace functions in bpf/user/agent.h.  This CL only has
schedghostidle, but we can extend the tracers[] easily.

The BPF objects and skeletons are pretty simple, and the bulk of any of the BPF
tools' C file is just boilerplate and output formatting.  For schedghostidle,
it's sgi_ functions are the guts of schedghostidle.c, adjusted slightly to run
in an agent.  e.g. no signal handling, start/stop time, etc.

Tested: agent + simple_exp + kill -USR1.  Waited 60 seconds to confirm the reset worked.