cl/415585725: ghost: fix up schedlat struct hist alignment
We don't need 64 byte alignment.  Internally, each element of the percpu array
is essentially a per-cpu pointer, so padding by cachelines doesn't help.

But we do need the 8 byte alignment.  The kernel rounds-up the size of the map
elements to 8.  If the compiler doesn't know about this, it'll think a struct
hist is e.g. 28 bytes instead of 32, and when we bpf_map_lookup_elem (which
returns an *array* of struct hist), the kernel will try to copy out 32 byte
structs, while the userspace is expecting 28 byte structs.

Tested: schedlat with and without alignment.  Without, we get memory corruption
in print_hists.