cl/417043353: ghOSt: Add SerializeVector/DeserializeVector
Also change `Deserialize<T>(size_t size)` to have a pointer to `T` as a parameter. If you are deserializing an array, it is hard to allocate an array of type `T` on the stack. This is because the array size is part of the array type, so you need to know the array size at compile time in order to use `Deserialize` as it is before this CL. If you do not know the array size until runtime, then the updated Deserialize in this CL works since you can pass an array pointer.