cl/423444735: ghOSt: Fix open source compile error caused by missing NULL macro for BPF
This macro is not available in libbpf/bpf_helpers.h in Linux until 5.13. Thus, klockstat.bpf.c fails to compile in the open source userspace repo because the open source ghOSt kernel is 5.11 (but does build internally because the NULL macro was backported to prodkernel, and the NULL macro was then copied to //third_party/linux_tools). Define this macro in the open source userspace repo until the open source ghOSt kernel is 5.13 or newer.

Reported by: hannahpan@