cl/429147376: ghOSt: Rename CompletelyFairScheduler to ThreadWait
We want to offer an option to run RocksDB in ghOSt without the PrioTable (and perhaps ditch the PrioTable altogether). Currently, RocksDB only uses the PrioTable to tell ghOSt which worker threads are runnable. This can be communicated via a futex instead of the PrioTable. Thus, we want to reuse the thread wait code currently only used by the CFS experiment. Thus, rename CompletelyFairScheduler to ThreadWait to reflect that the code will be used by both ghOSt experiments (when not using the PrioTable) and CFS experiments.