cl/419752488: Support for multiple ABI versions.
go/kcl/435202 advertises support for multiple ABI versions in /sys/fs/ghost/version.

Modify userspace to validate that the kernel has support for the ABI that the
agent is compiled with. Also request a specific ABI version when creating an
enclave.

Submitted in conjunction with go/kcl/435202

TESTED=all unit tests pass in virtme