cl/419752622: Calculate number of bits in a 'gtid' for 'pid' and 'seqnum' at runtime.
It was a mistake to export the layout of a 'gtid' in an ABI-specific manner
since it really depends on the underlying kernel PID_MAX_LIMIT. This limit
can change across kernel rebase and invalidate any older ABIs (definitely
not something we want).

The kernel exports PID_MAX_LIMIT as /proc/sys/kernel/pid_max_max which we
use to initialize GHOST_TID_SEQNUM_BITS at runtime. If we find ourselves
running on a kernel that predates 'sysctl kernel.pid_max_max' then assume
a default value of 0x400000.

Submitted in conjunction with go/kcl/438926

TESTED:
- all unit tests pass in virtme
- change PID_MAX_MAX to 0x100000 and verify that 'xx' and 'yy' are correct.