cl/420306805: ghOSt: Fix open source build error
See https://github.com/google/ghost-userspace/issues/3.

When compiling externally on a machine running an open source ghOSt kernel, the definitions in bpf/user/agent.h cause the compilation to fail because the definitions for the same types also exist in the Linux headers installed on the machine (so the definitions in bpf/user/agent.h are *redefinitions*).

Thus, to fix this, we add the GHOST_BPF macro to include/uapi/linux/bpf.h in the ghOSt kernel. We check for this macro in the userspace code. If the macro exists, we do not redefine the ghOSt kernel BPF types. Otherwise, we do.

This CL will be submitted concurrently with kcl/454355.