cl/420896884: Fix initialization-order-fiasco for GHOST_TID_SEQNUM_BITS.
Specifically the bug is that we are depending on the initialization of
GHOST_TID_SEQNUM_BITS (in base.cc) during the initialization of another
global Agent::kVersionCheck (in agent.cc). Since there is no ordering
across compilation units this can result in accessing GHOST_TID_SEQNUM_BITS
before it has been initialized.

READ of size 4 at 0x7ff981751180 thread T0
    #0 0x7ff98134368f in ghost::gtid(long) third_party/ghost/lib/base.cc:122:19
    #1 0x7ff98134196d in GetGtid third_party/ghost/lib/base.cc:212:28
    #2 0x7ff98134196d in Current third_party/ghost/lib/base.h:164:40
    #3 0x7ff98134196d in ghost::Exit(int) third_party/ghost/lib/base.cc:285:28
    #4 0x7ff981b93db4 in ghost::Ghost::MountGhostfs() third_party/ghost/lib/ghost.cc:147:5
    #5 0x7ff981b93fe5 in ghost::Ghost::GetSupportedVersions(std::__u::vector<unsigned int, std::__u::allocator<unsigned int> >&) third_party/ghost/lib/ghost.cc:155:5
    #6 0x7ff982102744 in ghost::Ghost::CheckVersion() third_party/ghost/lib/ghost.h:274:5
    #7 0x7ff98217c564 in __cxx_global_var_init third_party/ghost/lib/agent.cc:102:35
    #8 0x7ff98217c564 in _GLOBAL__sub_I_agent.cc third_party/ghost/lib/agent.cc
    #9 0x7ff9cc37b4cc in call_init (/usr/grte/v5/lib64/ld-linux-x86-64.so.2+0x1c4cc)
    #10 0x7ff9cc37b329 in _dl_init (/usr/grte/v5/lib64/ld-linux-x86-64.so.2+0x1c329)

Fix this by caching the result in a function-local static variable in
get_tid_seqnum_bits() which is guaranteed to be initialized the first time
it is accessed, and this initialization will happen before any other access.

TESTED=all units tests pass in virtme