cl/421617713: ghOSt: add extra param to timerfd payload
This adds an extra field, 'type' to the ghost_timer_fd structure. This allows the agent to distinguish between different types of timers that it may queue. The cookie field in the timer_fd then acts as an argument to be interpreted based on the type of timer.

Paired with kcl/445401.