cl/428685576: Fix flaky ApiTest.LocalAgentWakeup
There were two symptoms reported in b/214648944:
- unexpected state after commit: -2147483646 (GHOST_TXN_TARGET_ONCPU)
- test timed out

The test depends on a carefully orchestrated dance between two agents
scheduling a task on their sibling cpu (via sync_group_commit) and
then verifying that the sibling was woken up when the task scheduled
on the sibling cpu (thus testing the "local" agent wakeup).

This mostly works but breaks down when an agent wakes out of turn and
can happen in the following situations:
- after delivering a MSG_CPU_TICK.
- as a side-effect of CONFIG_QUEUE_WAKEUP.
- via set_curr_ghost (e.g. when an oncpu task enters ghost on one of the
  cpus in the enclave).
- MSG_TASK_PREEMPT generated on the sibling even though the overall
  sync_group commit failed (see go/kcl/468691).

There is also an inherent race in agent startup where sibling 1 can race
with the ping to sibling 0 that kicks off the test.

After trying _very_ hard to eliminate the spurious wakeups accept defeat
and yield the agent that wakes up out-of-turn. This dilutes the value of
the test since we now tolerate out-of-turn wakeups so verify that the
number of bogus wakeups is less than 5% of the legitimate wakeups.

The timeout issue is fixed by setting RTLA_ON_IDLE if the agent is
terminating on the local transaction.