cl/428685904: ghOSt: Fix PickNextGlobalCPU() race in SOL scheduler
Currently, when we pick a new CPU to run the global agent on, we preempt the currently running ghOSt task there. We then just assume the task is runnable and add it to the runqueue again. However, this preemption attempt could race with a TASK_BLOCKED, TASK_YIELD, etc.

The problem is the following: Since the task is on the runqueue, the SOL scheduler may try to schedule it again in some txn B. However, when the SOL scheduler ends up receiving that TASK_BLOCKED (or TASK_YIELD, etc.), it will end up calling SyncTaskState(), which asserts that SyncCpuState() is true. However, SyncCpuState() will return false since txn B failed due to a stale barrier. Thus, the assertion fails and the scheduler crashes.

We resolve this race by just pinging the remote agent to wake up in PickNextGlobalCPU() without attempting a manual preemption of the task. When the agent wakes up, it automatically preempts the ghOSt task that is running on its CPU and the kernel generates a TASK_PREEMPT message for that task. Thus, the SOL scheduler will receive a TASK_PREEMPT message for the preempted task and update the task's state accordingly. If it turns out that the task actually blocked, yielded, etc. (in other words, the task was not preempted by the agent), a TASK_BLOCKED/TASK_YIELD/etc. message will be generated instead, so the SOL scheduler will still update the task's state properly in those cases (while still avoiding the race).

Tested: for i in {1..100}; do ./simple_exp; done. Note that I disabled the task departure sub-tests in simple_exp since SOL does not support TASK_DEPARTED.