cl/428811065: ghost: handle race with /proc/PID/fd/
The files in fd/ can be closed while we are iterating through the directory
entries.  All we have is a directory entry, but it can disappear at any moment.
Previously, we'd check-fail on is_symlink(f).  (ENOENT).  Even if we passed
that check (stat), readlink could fail if the file was closed after the is_symlink
check.

Even if we find the memfd that matches the needle, don't crash if we fail to
open it.  The task could be exiting or something.

Tested: agent_exp and simple_exp.  simple_exp doesn't have the memfd we are
looking for, and for unrelated reasons had files opening and closing.