cl/428950798: Fix CapabilitiesTest.AgentNoNice
This test relied on a specific order of checks such that it could
get away with passing bogus values of 'ctl_fd' and 'queue_fd' to
SchedAgentEnterGhost(). This broke after cl/428809087.

Fix the test by creating an enclave and a channel and passing proper
file descriptors to SchedAgentEnterGhost() so we only test for the
lack of CAP_SYS_NICE.

Reported by: jhumphri@

[ RUN      ] CapabilitiesTest.AgentNoNice
third_party/ghost/tests/capabilities_test.cc:153: Failure
Value of: (*__errno_location ())
Expected: is equal to 1
  Actual: 9 (of type int)
[  FAILED  ] CapabilitiesTest.AgentNoNice (2 ms)
[----------] 4 tests from CapabilitiesTest (82 ms total)

[----------] Global test environment tear-down
[==========] 4 tests from 1 test suite ran. (84 ms total)
[  PASSED  ] 3 tests.
[  FAILED  ] 1 test, listed below:
[  FAILED  ] CapabilitiesTest.AgentNoNice

TESTED=CapabilitiesTest.AgentNoNice is successful after this change.