cl/429134192: ghost: add a helper to disable the ability to load bpf programs
Once we've inserted the bpf programs (attached or linked), we no longer need to
load or attach/link more programs.

In the off chance our agent is compromised, by disabling bpf_prog_load, the
agent would be unable to load new hostile programs that it could attach to
other places in the kernel.  Our agents typically run with CAP_BPF, though not
necessarily CAP_PERFMON, so the list of bpf programs we could load and attach
is limited.  This helper limits it further.

The existing programs we already loaded are considered "known good".  They
cannot be attached to random places in the kernel: only to BPF-PNT and BPF-MSG.

This helper is a one-way operation for the calling process.  A new agent that
attaches to the enclave will be able to load and attach new programs.  (Note
you cannot have more than one program attached per enclave, per attach point.
That's why we handle EBUSY in InsertBpfPrograms().

To be submitted with kcl/470253.