cl/433596269: ghost: deflake enclave_test DestroyAndSetsched()
Yet another error case: ENOENT.  I think this was from openat(), though without
an strace, I don't know for sure.

There are lots of ways this stuff can fail, so the errno list we have is not
exhaustive.  For the few cases that slip through the cracks, let's just ignore
them.  We have 100 loops to catch the race.

Arguably we can just remove this test too, since so far all failures have been
in the test, not the kernel.

Tested: enclave_test