ghost: initial iteration of a cfs agent
Currently, the agent is very simplistic, it assigns new tasks in a round-robin fashion to CPUs. Each CPU has a runqueue; when ghost receives a message to schedule a ghost task on a cpu, it simply plucks the one with the lowest vruntime. This plucked task will run until preempted by a task from another sched class, never by ghost itself as we do not list for sched_tick events, meaning that once we pick a task to run, it will run until it blocks itself (futex, io, pagefault, etc) or another sched class preempts it.

I don't know if abusing std::multiset is the way to go to get a red-black tree ¯\_(ツ)_/¯. I ran into annoying issues with std::set that took a while to debug because I was unsure how to instrument the its underlying functions due to the heavy use of templates.

Below is a (non-exhaustive) list of steps needed to bring this agent to parity with the in-kernel cfs:
- add a sched_latency knob/listen for sched_tick: right now tasks will run until they voluntarily yield (up to preemption by another sched class)
- load balancing
- nice values
- work stealing
- group scheduling