ghOSt: Use absl::Barrier and -std=c++2a for gcc9 compatibility
This change updates gcc-9 incompatible <latch> and -std=c++20 to their
compatible alternatives: absl::Barrier and -std=c++2a. This patch will
allow users to compile and try out ghOSt as instructed in README.md.

The current compilation/installation steps described in README.md are
based on gcc-9 which is the default in Ubuntu 20.04. A couple of recent
changes (introduction of <latch> and -std=c++2a) are not supported by
gcc-9 and required users to manually install gcc-10 or higher or update
the code by themselves.