ghost: add bpf_prog_load capabilities_test
Interesting things related to process-global state:

- the AgentNice test was calling Enclave::Ready(), which disables the
ability to load bpf programs.  Since this test was running enclaves from
within a single process, it would lose the ability for future tests to load
bpf programs.  If you make an enclave and ready it, do so in a fresh
process.

- The agent_bpf registration stuff was torn down when a FullAgent exited in
LocalEnclave::PrepareToExit(), which happens before the enclave dtor.  This
is an optimization, since we want to start closing those FDs ASAP during a
live update.  However, if you make an enclave without a FullAgent (which is
different than an AgentProcess), then we weren't closing these FDs, and we
were leaving blobs around in the process-global agent_bpf_ registry (in
agent_bpf.c).  The fix is to make PrepareToExit() idempotent.  That way we
can call it early when the FullAgent shuts down and again when the
LocalEnclave destructs to ensure it happens.