ghost: cfs: massage code to look more like in kernel CFS
This should make it easier to replicate more complicated in kernel logic
like load balancing and some locking semantics surrounding run queues.

Now, when processing messages, we have two case:
- the task associated with the message is cs->current
- the task associated with the message is not cs->current

In the former case, we update the task's state, and defer freeing it or
adding it back to the rq until PickNextTask. IOW, PickNextTask is now
the sync point for cs->current, much like in the kernel. In the latter
case, we just do the operation like normal and possibly free it or place
it on the runqueue.

A side effect of this practice is that we no longer have to store an
extra pointer for the purpose of checking if the previously oncpu task
should still be oncpu post preemption since we just keep it in
cs->current.

An astute reader will notice that it is possible from req->Commit() to
fail. Since we set cs->current before req->Commit() it is possible that
cs->current will not be oncpu. This is okay because we will go through
another Scheduling loop which will update the state for any tasks that
are not cs->current and, possibly, update the state for cs->current,
which will then be all sync'ed in PickNextTask again.

Additionally, this change fixes a bug in our custom Less function, where
tasks with the same vruntime would be considered equal, causing
incorrect erasures.

Additionally, this CL adds the following debugging utilities:
- DSTREAM(level), DCFS_PRINT(level)
- A state machine validator