ghost: add FdServer
FdServer is a server that passes an FD over a unix domain socket.
Agents can use this to pass a data structure (e.g. a bpf array
map) to an application instead of hunting around in /proc for
magic FDs (e.g. OpenGhostShmemFd).

The client needs to get the path to the socket and the nonce via some
other mechanism, such as gRPC or even the command line.

By default, FdServer serves the FD once and shutsdown.  You can set it
to serve forever too (and technically "nonce" is no longer a nonce!).
The "serve forever" approach is more analogous to Plan 9's #srv
device[1], if you're curious.

One thing you could do with "serve_forever" is have the agent run an
AsyncFdServer with serve_forever=true, then put the port and nonce in a
file in /tmp that the app can find.  That sidesteps the whole gRPC
thing, and might help in prototyping.

Tested: fd_server_test, and manually used fdcat and fdsrv:

$ echo hello > fdsrv NONCE &
@074e0 NONCE
$ fdcat @074e0 NONCE
hello

$ echo hi > /tmp/foo
$ fdsrv NONCE < /tmp/foo &
@36911 NONCE
$ fdcat @36911 NONCE
hi

[1] https://github.com/brho/akaros/blob/master/kern/drivers/dev/srv.c