Fix occasional hang in ApiTest.RunDeadTask
pthread_join() strikes again!

In this particular case t.join() returned "early" before the
thread had fully exited (in fact it had gotten offcpu via a
_cond_resched()). This raced with the AgentProcess destructor
calling TerminateAgentTasks(). When the satellite agent had
exited there was no cpu on which the 't' could schedule even
though it was runnable.

Fix this by exiting the spinning agent immediately after
Finished() is asserted. If there are tasks remaining in
the enclave they will be moved into the fair sched_class
by the kernel when the enclave is destroyed.

Tested: 10000 iterations of ApiTest.RunDeadTask