ghost: reject AgentRPCs if the enclave is not online
Needed for live updating an agent with a stubby/gRPC server, so the client can
tell (via a failed RPC) that the agent is offline.

Here's the scenario:
1) The Agent Process (AP) *parent* runs the Stubby server.  Why?  Because it's
a nightmare to put it in the AP due to the ForkedProcess.  Besides, part of the
reason for the whole AP stuff was to put stuff like a Stubby server in the
parent.

2) The parent makes the Stubby server *after* the AP was created.  AP creation
already handles live-update: if you try to create a second AP, it'll block
until the first one closes agent_online and kills its agent tasks.

3) Once the parent kills the AP (which closes agent_online), it'll kill the
Stubby server.  There's no guarantee the parent could kill the stubby server
first: the AP could crash (which closes agent_online, eventually).

4) Eventually, the parent closes the stubby server.

Meanwhile:
5) A second agent (parent + AP) is waiting to take over the enclave.  Once its
AP is created, the first agent has closed agent_online and has vacated.

6) Then the second agent starts its stubby server on the same port (in a future
CL).  It'll retry until the port/path/socket is available.

7) The client is watching agent_online and will know when the AP closes it.

So there's a window of time after agent_online closed, but before the first
stubby server is gone and before the second is up.  The client sees
agent_online closed, and the immediately tries to contact the stubby server,
making RPCs.

However, it's still talking to the old server!  Now, since the AP is dead, the
AP's rpc handler probably won't return anything sensible, and likely won't even
be around to respond to the Notification (internal Agent RPC signalling).
That's actually OK: when the server is killed, the Stubby RPC will die too,
which the client can handle.

The thing that the client can't handle is getting a *correct* response from the
*wrong* agent: the previous one.  That could happen if agent_online is closed,
but the RPC server is running: picture a long delay in
LocalEnclave::PrepareToExit(), right after closing the FD, but before
returning.  That's in the window.

With this change, any Agent RPCs will fail with -ENODEV if the enclave is not
online, which happens before closing agent_online_fd_.

Tested: enclave_test, edf_test, api_test