ghOSt: Fix bug in EnclaveTest.CpuListComma
This test creates a CpuList from MachineTopology() that contains CPUs
that may not exist in the machine. This is okay for now but could
trigger DCHECKs and access out-of-bounds memory in the future if we call
different methods on the CpuList or if the implementation of CpuList
changes. Thus, we instead use TestTopology() since the test topology has
enough CPUs for the test.

TESTED=enclave_test

PiperOrigin-RevId: 470156222