Rework 'txn->state' expectations for a sync_group commit.
In conjunction with go/kcl/536138 we are changing the semantics of
a sync_group commit as it relates to the 'txn->state' of the cpus
involved in the commit.

Prior to this change a failed sync_group commit was expected to
have 'txn->state' indicating failure in all transactions involved
in the group.

This is now changing such that a failed sync_group will have at
least one txn->state representing a failed transaction. The rest
may indicate that they were successful or not attempted at all.

Tested: all unit tests pass in virtme
PiperOrigin-RevId: 471290116