Initialize members of 'struct Runner' at declaration.
Without this all three members were uninitialized which led to test failures
when compiled with "--dynamic_mode=off" in virtme (but not with opt or dbg):

[ RUN      ] AgentTest.AgentBlock
RpcHandler: runner_.ok = -85
third_party/ghost/tests/agent_test.cc:964: Failure
Expected equality of these values:
  ap.Rpc(kRunBlockTest)
    Which is: -86
  0
[  FAILED  ] AgentTest.AgentBlock (81 ms)

The test was failing consistently with the same error (-86) which makes
sense once you do an objdump of RpcHandler. The compiler was translating
this snippet of code:

        if (runner_.ok) {
          response.response_code = 0;
        } else {
          response.response_code = -1;
        }

into the following assembly:
  	movsbq 0x8202(%rdi),%rax	<<< rax = *(int64_t *)&runner_.ok
  	dec    %rax
  	mov    %rax,(%rcx)		<<< response_code = --%rax

which makes sense (ok=1,response_code=0 and ok=0,response_code=-1)
except when Runner::ran_func is uninitialized (to a non-zero value)
which makes the RpcHandler sample Runner::ok before the agent has
a chance to initialize it!

PiperOrigin-RevId: 471902476