ghost: biff: don't call bpf_printk() all the time
bpf_printk() is really bpf_trace_printk(), which requires CAP_PERFMON (or
SYS_ADMIN).  Not all of our agents run as root, so this is a minor issue.

For now, if you're debugging and have CAP_PERFMON, you can toggle the macro
definition of bpf_printd() and recompile.

Alternatives for the future:
- autodetect from userspace if we have CAP_PERFMON and switch these calls off.
  (requires libbpf support).
- add a ghost print helper, essentially bypassing CAP_PERFMON.
- don't use printk: send messages over a ringbuffer map to userspace for these
  error cases.
- have some "oh crap things went to hell" helper that prints and kills the
  enclave.

Tested: biff_test, with the macro toggled each way
PiperOrigin-RevId: 477786640