ghOSt: Relax failure conditions for EnclaveTest.CpuListComma
This change relaxes EnclaveTest's CpuListComma test case to make it pass on
machines with 32 or less CPUs (i.e. GCP Ubuntu E2's). Previously, the test
checked whether `WriteEnclaveCpus` succeeds and, if it fails, CHECKs whether
`errno` is not `EOVERFLOW`. However, if the machine has 32 or less CPUs,
`WriteEnclaveCpus` can also legitimately return `EOVERFLOW` if the number of
CPU mask bits (`nr_cpumask_bits`) is set based on the actual number of CPUs.

To make the test case pass, this CL:
* EXPECTs `WriteEnclaveCpus` to succeed if the machine has more than 32 CPUs.
* If the machine has 32 or less CPUs, we still try `WriteEnclaveCpus`, but
  expect it to return `EOVERFLOW` and otherwise mark the case as failure.

PiperOrigin-RevId: 482120819