ghOSt: Respect new ghOSt task's affinity set before entering
This change updates ghOSt CFS agent to respect a task's affinity which was set
before it enters ghOSt. When the agent receives a `TASK_NEW` message, we make
`sched_getaffinity` syscall to get the task's initial task affinity. In an
unlikely case where agent fails to get the initial affinity, the agent falls
back to have the task take any CPU in the enclave and be scheduled.

Major Changes:
* Updated `CfsScheduler::TaskNew` to look up a task's affinity via
  `sched_getaffinity` when it first shows up.
* Implemented a new unit test to `cfs_test` to test whether the initial
  affinity is respected.

PiperOrigin-RevId: 483522385