ghost: clear SetDeliverCpuAvailability during teardown
During a clean (non-crashing) teardown, BPF programs are removed before the
agent tasks are killed.  TerminateAgentTasks() calls enclave PrepareToExit() to
detach the BPF programs before terminating the actual agents, since the agents
take a long time (relatively speaking) to exit.

After the programs are removed, but before the agents exit, the CPU may go
through an availability edge.  Prior to this change, the enclave was still set
to deliver availability messages, but bpf-msg wasn't running anymore.  That
means that we were trying to send those messages to userspace.

That isn't a big deal, but if the agent has a very small channel, such as sgs's
channel of size 1, you'll overflow the queue and get a warning or notice from
the kernel.

PiperOrigin-RevId: 487046900