ghOSt: Remove CHECK statements from RPC buffer serialization
The non-static_assert CHECKs can easily be triggered at runtime if too large of
a type is supplied. This is pretty trivial, for example, with std::string's,
since they can be of arbitrary length.

Replace the CHECKs with absl::Status and absl::StatusOr return values. This
lets code that wants to harden against failures do so by handling the error
rather than CHECK'ing on it.

No functional change as a result of this patch.

PiperOrigin-RevId: 495992876