ghOSt: Move per-cpu channel association forward in agent threadbody
Instead of binding all the agents to the scheduler's default channel at
the beginning of the agent ThreadBody, this change opens up a new path
where scheduler can try to bind the right channel for that agent from
the beginning.

To note, this change does not force each scheduler to decide this from
the beginning. If a scheduler would like to decide which channel to bind
for an agent later, it can return default channel as the return value
for `GetAgentChannel` function, which is the current and will be the
default behavior after this CL.

PiperOrigin-RevId: 501392706