ghOSt CFS: Rename Rq::Erase* to Rq:Dequeue*
In ghOSt CFS we use a C++ `set` as RQ storage since it's implemented
using an RB-Tree. To remove an entry from the RB-Tree `set::Erase()` is
used. However, from a CFS perspective, this operation is about dequeuing
a task.

Make ghOSt CFS code clear by `s/Erase/Dequeue/` every RQ operation
involving the dequeuing of a task. This also avoids the confusion around
possibly "erasing" task's data, which is an operation we expect instead
only when a task is dead.

No Functional changes.

PiperOrigin-RevId: 501867920