ghOSt CFS: Better encapsulate RQ's empty checks
The RQs APIs have an `Empty()` method to check for tasks in the RQ.
Always use this API to keep encapsulated the implementation details.
Rename it to `IsEmpty()` to better call out this is a check.
Use the set's `empty()` method, instead of comparing with its size, to
check for the RQ being empty.

No Functional changes.

PiperOrigin-RevId: 502371955