ghost: add MSG_TASK_LATCH_FAILURE
Used only by BPF to tell the agent that a task failed to latch, with the
guarantee that the task's lock is held during the handler.  This
prevents other message handlers from running for the same task
concurrently.

For a similar guarantee, userspace can check the transaction region from
the agent cpu to which the task is associated.  Agent code from the cpu
that handles the associated channel gets the same serialization as BPF's
task lock.  See the flux Treatise On Locking in an upcoming CL for more
info.

PiperOrigin-RevId: 504418972