ghost: add must_resched to BPF-PNT's context
This lets BPF-PNT force a task off cpu.

One use: if the task you want is EBUSY or otherwise unable to run, we
need to context switch to the idle task, then restart PNT.  However, if
prev isn't blocking or yielding, it will stay on the cpu, and idle won't
get picked.

PiperOrigin-RevId: 504419051