ghOSt CFS: Refactor idle balance to be more upstream aligned
Add a CpuIdleType and use it to track the type of idle load balance
requested via the standard LoadBalanceEnv.

Make it clear we go the NewlyIdleBalance() path when PickNextTask() does
not return a task. This is indeed the only condition we have so fare
where we know that a running CPU is going to be idle and thus it makes
sense to trigger a task stealing balance from a possible busier CPU.

Do that by renaming IdleLoadBalance() into NewlyIdleBalance(), which
makes the name better matching its description and the behavior
implemented in the only call site we use it. This also allows to embed
balance information (e.g. being newly idle) into load balance related
functions and the corresponding env.

Use LoadBalanceEnv as much as possible. Thus, pass it in as the only
parameter for the ShouldWeBalance() call. This also better align with
upstream and make that function ready for future additions.

No Functional changes.

PiperOrigin-RevId: 505077898