ghOSt CFS: Move state changes due to setaffinity to TaskPreempted/TaskYield
This change moves the logic to kick a task off CPU in
TaskAffinityChanged message to TaskPreempted. It also adds affinity
handling logic to TaskYield so that we consistently examine affinity
when a task is getting off-cpu.

The previous implementation did so in TaskAffinityChanged, assuming that
TaskAffinityChanged message is always followed by TaskPreempted message.
However, we can still have TaskAffinityChanged messages without
TaskPreempted messages following them (i.e., getting TaskBlocked or
TaskYield instead) and it is not an unintended behavior.

PiperOrigin-RevId: 505187756