ghost: add the flux scheduler
Flux is the name of the framework and it's example BPF scheduler for
preemptive hierarchical scheduling.

The idea is to make a scheduler out of cooperating mini-schedulers or
sub-schedulers, in a hierarchy.  Schedulers of schedulers allocate cpus
to other schedulers.  Schedulers of threads run threads on cpus.

By decomposing a bigger scheduler into subschedulers, you can specialize
cpus to run classes of threads, without worrying about one big scheduler
that can handle all cases.  For example, you could have a high-priority
scheduler for certain tasks with a custom policy, and then use the
remaining cpus with a CFS-like scheduling policy for lower priority
threads.

The flux model isn't particular to BPF, or even ghost.  It's based off
of the Lithe userlevel threading framework, which involved dynamically
adding and removing schedulers from the hierarchy.  Think of a parallel
for loop call from OpenMP spawning a mini scheduler for the life of the
for loop.

In this scheduler, there is a single root scheduler name "roci" (Root
One-Child and Idle)  that allocates cpus to two child schedulers: "biff"
and "idle".  Biff has the policy we all know and love: global fifo.
Idle simply lets a cpu go idle.

To understand the interaction between schedulers, consider the case of a
thread (belonging to Biff) waking up.  Biff now needs more cpus (based
on nr_runnable + nr_running), and it requests cpus from Roci.  Roci will
preempt them from Idle and allocate them to Biff.  When the kernel
temporarily takes a cpu away from ghost (e.g. to run CFS), the cpu will
be preempted from all schedulers.  So if a cpu is allocated to Roci,
which allocates it to Biff, and is then preempted by the kernel (ghost
MSG_CPU_BUSY), Roci responds to the preemption by trying to take another
cpu from Idle and give it to Biff.

Check out flux.bpf.c to see how the overall agent_flux scheduler is
composed of roci, biff, and idle.  Arguably, roci is rather particular
to agent_flux, but split it out as its own scheduler in case we want
to use it in other agents.

PiperOrigin-RevId: 506057755